UBUNTU: linux-tools -- pull out the perf binary into a binary package
We are getting complex dependancies on the linux-image package as a result
of including the perf binary in the package.  This ties us to the versions
of binutils and libc on the build release, which limits the installability
of that kernel on older releases.  This is a problem for testing.

Move the binary over to a release and architecture specific binary package
linux-tools-VERSION-ABI.  Rename the current linux-tools package over to
linux-tools-common and depend the binary packages on it.  We will also
need to add a new linux-tools meta package to the linux-meta package.

Signed-off-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>