UBUNTU: SAUCE: Disable function tracing after hitting __schedule_bug
If we hit __schedule_bug, then something bad has happened, and it would
be helpful to get an accurate function trace to see what's going wrong.
The debugfs function tracer works great for this task, but it spits out
a ton of information both before and after the real bug occurs. This
change stops the tracing when we enter __schedule_bug, so the tracing
report is smaller and leaves the developer with the needed information
right at the end of the trace.

This should not cause a regression or any negative effects unless
someone truly wanted to trace through a __schedule_bug call. In that
case, they can recompile their own kernel without this call. The vast
majority of tracings that hit __schedule_bug would be enhanced by this
change.

To see how tracing with this change can be used for "scheduling while
atomic" bugs, see http://bugs.launchpad.net/bugs/534549.

Signed-off-by: Chase Douglas <chase.douglas@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Acked-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>