UBUNTU: config -- allow locally specified configuration overrides
When updating configuration you often want to bulk update specific options
in every configuration.  For new options this is trivial as you can simply
add them to config.common.<family>.  For existing options you need to
either remove the options globally and then add them, or add them to
every flavour file.  This is particularly problematic for mainline builds.

Add support for a new $(DEBIAN)/config/OVERRIDES file which if present is
appended to the built configs before processing, providing an absolute
override for all configurations.  This is not expected to be present in
the normal case.

Signed-off-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Colin King <colin.king@canonical.com>