UBUNTU: [Config] enable MMIOTRACE for graphics debugging
Enable MMIOTRACE to allow us to trace low level card operations.  This is
commonly used to debug graphics issues particularly under Nouveau.

The cost is near-zero for having the MMIOTRACE option enabled with the
trace collection turned off (the default).  We take one additional branch
in the mmio map/unmap paths which are non-critical initialisation/teardone
paths in normal usage.  The actual mmio operations are not affected at all.
When we enable trace collection the overhead is high, taking a page fault
for each access to the mmio mapping.

Signed-off-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Colin King <colin.king@canonical.com>